fix : review reflection - add constructor of dto, move dto to port package
