+++ piikii-application/src/main/kotlin/com/piikii/application/port/input/room/dto/response/RoomSaveResponseForm.kt
@@ -1,0 +1,12 @@
+package com.piikii.application.port.input.room.dto.response
+
+import com.piikii.application.domain.room.Room
+import java.util.UUID
+
+class RoomSaveResponseForm(
+    private val roomId: UUID,
+) {
+    constructor(room: Room) : this(
+        roomId = room.roomId!!,
+    )
+}
